==> Customer Personality Analysis/notebooks/04-Preprocessing.ipynb <==
### Introduction

This notebook focuses on the preprocessing steps required to prepare the dataset for machine learning models. The dataset contains both categorical and numerical variables, which require specific transformations to ensure compatibility with algorithms and to improve model performance. 

The key preprocessing steps covered include:

- **Handling of categorical variables** using One-Hot Encoding (OHE) to convert them into a numerical format.
- **Scaling of numerical variables** using MinMaxScaler to ensure they are within the same range.

These transformations are essential for ensuring that the dataset is ready for further analysis and modeling.



import pandas as pd
import numpy as np



%run ../customer_personality_analysis/utils/pandas_missing_handler.py
%run ../customer_personality_analysis/utils/pandas_explorer.py

## Data Load and First Visualization:

path = '../customer_personality_analysis/data/cleaned_data.csv'
df = pd.read_csv(path).drop(columns=['Unnamed: 0'])
df.head()

## Creating new features

# Creating a function to labelice features
def categorice_feature(data, feature, labels):
    if len(labels) == 3:
        Q1,Q3 = data[feature].quantile([0.25, 0.75])
        # Applying labels
        categories = data[feature].apply(lambda x: labels[0] if x <= Q1 else (labels[1] if x <= Q3 else labels[2]))
    elif len(labels) == 4:
        Q1, Q2, Q3 = data[feature].quantile([0.25, 0.50, 0.75])
        categories = data[feature].apply(lambda x : labels[0] if x <= Q1 else (labels[1] if x <= Q2 else labels[2] if x <= Q3 else labels[3]))
    return categories

# Creating 'Age_level' feature
labels = [1,2,3]

df['Age_level'] = categorice_feature(
    data = df,
    feature = 'Age',
    labels=labels,
)

# Creating 'Income_level' feature based on 'Income'
Income_level = [1,2,3,4]
df['Income_level'] = categorice_feature(
    data=df,
    feature='Income',
    labels=Income_level
)

# Creating 'Tenure_level' feature
tenure_levels = [1,2,3,4]
df['Tenure_level'] = categorice_feature(
    data=df,
    feature='Tenure_Months',
    labels=tenure_levels
)

## Features Selection

selected_features = ['Education','Marital_Status','Income_level','Kidhome','Teenhome','Tenure_level','Age_level']

- Feature selection for demographic segmentation was performed, considering only demographic features.

## Mapping Categorical Features to Numerical:

from sklearn.preprocessing import OrdinalEncoder

# Creating a copy of the dataset:
encoded_df = df[selected_features].copy()

# Initialize the OrdinalEncoder with categories defined manually
education_levels = ['Basic', '2n Cycle', 'Graduation', 'Master', 'PhD']
encoder = OrdinalEncoder(categories=[education_levels])

# Fit and transform the 'Education' column
encoded_df['Education'] = encoder.fit_transform(df[['Education']])

# Applying Ohe to Marital_Status feature
encoded_df = pd.get_dummies(columns=['Marital_Status'], data=encoded_df).drop(columns=["Marital_Status_Uncommitted"])
encoded_df['Marital_Status_Committed'] = encoded_df['Marital_Status_Committed'].astype(int)

# Display the first few rows
encoded_df.head()

## Standardizing Numerical Features:

from sklearn.preprocessing import MinMaxScaler

# Selecting numerical features:
numerical_features = encoded_df.select_dtypes(include=[np.number]).columns

# Initializing MinMaxScaler:
scaler = MinMaxScaler()

# Applying Standardization:
encoded_df[numerical_features] = pd.DataFrame(scaler.fit_transform(encoded_df[numerical_features]), columns=numerical_features)
encoded_df.head()

## Saving the Preprocessed Dataset in a csv file:

encoded_df.to_csv('../customer_personality_analysis/data/preprocessed.csv')

## Preprocessing Summary

In this preprocessing stage, the following actions were taken to prepare the dataset for machine learning models:

1. **One-Hot Encoding for categorical variables:**  
   Categorical variables were transformed using One-Hot Encoding (OHE), ensuring they were converted into binary columns, preserving their original information for model training.

2. **MinMaxScaler for numerical variables:**  
   Numerical variables were standardized using the MinMaxScaler method, which scales values to a [0, 1] range, ensuring all variables have a comparable scale.

These transformations ensure that the dataset is adequately prepared for further modeling steps.
