==> Customer Personality Analysis/notebooks/06-Clustering analysis.ipynb <==
# Cluster Analysis

Customer segmentation is a crucial step in understanding the unique characteristics and behaviors of different groups within a customer base. In this project, clustering was performed using **Principal Component Analysis (PCA)** for dimensionality reduction and **KMeans** for grouping customers into clusters. The resulting segmentation aims to uncover patterns and insights that can inform targeted strategies and decision-making.

This phase focuses on analyzing the clusters to understand their distinct features and distributions. By exploring the characteristics that define each group and differentiate them from others, we aim to:

- Identify key attributes that influence the formation of clusters.
- Interpret the patterns and preferences within each segment.
- Highlight differences across clusters to provide actionable insights.

The findings from this analysis will serve as a foundation for strategic business decisions, such as personalized marketing, product recommendations, and resource allocation.


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import sweetviz as sv

%run ../customer_personality_analysis/utils/pandas_explorer.py

path = '../customer_personality_analysis/data/clustered.csv'
df = pd.read_csv(path)
df.head()

df[''].value_counts()

## Selecting features used for clustering

categorical_features = ['Education','Marital_Status','Kidhome','Teenhome','Cmp_Accepted']
continuous_features = ['Tenure_Months','Age','Income','Total_bought','Total_purchases']

## Cluster 0 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 0][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 0][col].describe()}\n")


- **Education**: Predominantly well-educated; 51.21% Graduation, 24.57% PhD, 15.92% Master.
- **Marital Status**: Entirely "Commited."
- **Family Structure**:
  - **Kidhome**: 92.73% have no children, 7.27% have one.
  - **Teenhome**: 55.02% have no teenagers, 44.29% one, and 0.69% two.
- **Campaign Engagement (Cmp_Accepted)**: Low engagement, with 64.01% not accepting any campaigns; smaller percentages accept 1-4 campaigns.


- **Tenure**: Short, averaging 5.30 months (range: 0–11).
- **Age**: Middle-aged, mean 45.93 years (range: 20–71).
- **Income**: Financially stable, mean $71,356 (range: $49,605–$96,876).
- **Consumption**:
  - **Total Bought**: High, averaging 1037 (range: 42–2524).
  - **Total Purchases**: Medium, averaging 19.63 (range: 5–32).

### Key Insights
Highly educated, financially stable, and consumption-driven middle-aged individuals with minimal family responsibilities, moderate campaign engagement and relatively short tenure.



## Cluster 1 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 1][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 1][col].describe()}\n")


- **Education**: Highly educated; 47.79% Graduation, 24.10% PhD, 18.88% Master, 9.24% 2n Cycle.
- **Marital Status**: Entirely "Uncommitted."
- **Family Structure**:
  - **Kidhome**: 86.35% have no children, 12.85% one, and 0.80% two.
  - **Teenhome**: 57.03% have one teenager, 40.96% one, and 2.01% two.
- **Campaign Engagement (Cmp_Accepted)**: Low acceptance rates; 73.09% accepted none, smaller proportions accepted 1–4 campaigns.


- **Tenure**: Long tenure, averaging 17.19 months (range: 12–22).
- **Age**: Middle-aged, mean 48.97 years (range: 19–71).
- **Income**: Moderate, mean $63,439 (range: $22,304–$102,692).
- **Consumption**:
  - **Total Bought**: Medium, averaging 1021 (range: 18–2209).
  - **Total Purchases**: Medium, averaging 20.44 (range: 4–33).

### Key Insights
This cluster represents financially stable, middle-aged individuals with high education levels but "Uncommitted" marital status. They have longer tenures, moderate income, and significant consumption. Engagement with marketing campaigns remains low, suggesting room for tailored strategies.


## Cluster 2 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 2][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 2][col].describe()}\n")


- **Education**: Predominantly well-educated; 40.97% Graduation, 27.78% Master, 26.74% PhD. Few with 2n Cycle (4.17%) or Basic education (0.35%).
- **Marital Status**: Entirely "Committed."
- **Family Structure**:
  - **Kidhome**: 61.11% have no children, 36.46% one, and 2.43% two.
  - **Teenhome**: 84.03% have one teenager, 8.68% two, and 7.29% none.
- **Campaign Engagement (Cmp_Accepted)**: Very low; 85.42% accepted none, with minimal engagement in 1–3 campaigns.

- **Tenure**: Short tenure, averaging 10.07 months (range: 0–22).
- **Age**: Older demographic, mean 56.04 years (range: 38–71).
- **Income**: Modest, mean $48,110 (range: $4,428–$93,404).
- **Consumption**:
  - **Total Bought**: Low, mean 359.56 (range: 11–1616).
  - **Total Purchases**: Moderate frequency, mean 14.11 (range: 1–35).

### Key Insights
This cluster represents older, married individuals with modest incomes and shorter tenure. Despite their high education levels, consumption is relatively low, and engagement with marketing campaigns is minimal. Strategies could focus on increasing loyalty and targeting low-cost value propositions.


## Cluster 3 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 3][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 3][col].describe()}\n")


- **Education**: High proportion of highly educated individuals; 51.78% have a Graduation degree, 29.64% PhD, 14.23% Master, and 4.35% with 2n Cycle.
- **Marital Status**: Entirely "Uncommitted."
- **Family Structure**:
  - **Kidhome**: 77.47% have no children, 22.13% one, and 0.40% two.
  - **Teenhome**: Evenly distributed between households with one teenager (49.01%) and none (47.43%), with 3.56% having two teenagers.
- **Campaign Engagement (Cmp_Accepted)**: Moderate; 71.94% accepted none, while 18.18% accepted one, and smaller proportions accepted more.

- **Tenure**: Very short tenure, averaging 5.25 months (range: 0–11).
- **Age**: Diverse age range with a mean of 48.36 years (range: 19–74).
- **Income**: Relatively high, mean $63,369 (range: $7,144–$113,734).
- **Consumption**:
  - **Total Bought**: High, mean 785.77 (range: 11–2525).
  - **Total Purchases**: Frequent, mean 16.25 (range: 4–32).

### Key Insights
This cluster is characterized by uncommitted, educated individuals with short tenure and relatively high income. They demonstrate active purchasing behavior, both in terms of total amount and frequency, but moderate engagement with marketing campaigns. Efforts could focus on loyalty programs and campaigns that emphasize value for educated, transient individuals.



## Cluster 4 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 4][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 4][col].describe()}\n")


- **Education**: Balanced educational background:
  - 51.18% with Graduation, 15.75% PhD, 12.99% 2n Cycle, 11.02% Master, and 9.06% Basic.
- **Marital Status**: Entirely "Committed."
- **Family Structure**:
  - **Kidhome**: 75.98% have one child, 21.26% have none, and 2.76% have two children.
  - **Teenhome**: Predominantly no teenagers (72.83%), followed by one teenager (25.98%) and 1.18% with two teenagers.
- **Campaign Engagement (Cmp_Accepted)**: Low; 91.34% accepted none, while 8.27% accepted one, and 0.39% accepted two.

- **Tenure**: Long tenure, averaging 17.28 months (range: 12–22).
- **Age**: Younger demographic, mean age 37.80 years (range: 18–67).
- **Income**: Low to moderate, mean $31,897 (range: $2,447–$66,373).
- **Consumption**:
  - **Total Bought**: Low, mean 187.38 (range: 11–1,730).
  - **Total Purchases**: Moderate, mean 10.20 (range: 4–43).

### Key Insights
This cluster comprises younger, committed individuals with low to moderate income. They are characterized by long tenure and minimal engagement with campaigns. Purchasing behavior is modest in both frequency and total amount. Strategies for this group could focus on personalized offers to increase purchasing activity, particularly for families with one child and no teenagers.


## Cluster 5 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 5][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 5][col].describe()}\n")


- **Education**: This cluster has a high level of education:
  - 57.10% with a Graduation, 22.71% have a PhD, 12.93% have a Master's degree, and 7.26% completed the 2nd Cycle of education.
- **Marital Status**: All individuals in this cluster are "Committed."
- **Family Structure**:
  - **Kidhome**: The majority (89.91%) have no children, while 10.09% have one child.
  - **Teenhome**: Half (50.79%) have one teenager, while 47.63% have no teenagers, and 1.58% have two teenagers.
- **Campaign Engagement (Cmp_Accepted)**: Most individuals (67.51%) have not accepted any campaigns, with 19.87% accepting one, 7.26% accepting two, 4.42% accepting three, and 0.95% accepting four campaigns.

- **Tenure**: The average tenure is 17.59 months, with a range from 12 to 23 months.
- **Age**: The average age is 46.42 years, ranging from 20 to 73 years.
- **Income**: On average, income is $68,281, ranging from $35,797 to $105,471.
- **Consumption**:
  - **Total Bought**: The average number of items purchased is 1,179, ranging from 68 to 2,440.
  - **Total Purchases**: On average, 21.21 purchases were made, ranging from 7 to 39.

### Key Insights
This cluster represents individuals with medium and high education levels and average incomes. The majority of individuals have no children, but a significant portion has teenagers. They show a moderate level of engagement with campaigns, having accepted one or more offers. This group is characterized by an average consumption, with frequent purchases of a wide variety of products. Targeting this cluster for more personalized campaigns could further increase engagement, especially with products or services that align with their higher purchasing habits and family needs.



## Cluster 6 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 6][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 6][col].describe()}\n")


- **Education**: The education distribution in this cluster includes:
  - 48.92% with a Graduation, 16.19% with a Master's degree, 15.83% with a 2nd Cycle of education, 15.47% with a PhD, and 3.60% with a Basic level of education.
- **Marital Status**: All individuals are "Committed."
- **Family Structure**:
  - **Kidhome**: 82.37% of individuals in this cluster have one child, 11.15% have no children, and 6.47% have two children.
  - **Teenhome**: 73.38% have no teenagers, while 26.62% have one teenager.
- **Campaign Engagement (Cmp_Accepted)**: A significant majority (91.01%) have not accepted any campaigns, while 8.99% have accepted one campaign.

- **Tenure**: The average tenure is 5.4 months, with a range from 0 to 11 months.
- **Age**: The average age is 38.14 years, with a range from 18 to 55 years.
- **Income**: On average, income is $32,922, with a range from $4,023 to $71,427.
- **Consumption**:
  - **Total Bought**: The average number of items purchased is 98, ranging from 8 to 989.
  - **Total Purchases**: On average, 7.97 purchases were made, with a range from 4 to 25.

### Key Insights
This cluster is characterized by individuals with moderate to high levels of education. They are mostly young to middle-aged (average age of 38 years) and have a relatively short tenure (average of 5.4 months). Most individuals have one child and few engage in campaigns. Income levels are moderate, with an average income of $32,922. This group has a lower level of consumption compared to other groups, with an average of 98 items purchased. Targeting them with specific offers might be effective, especially with products that appeal to young families and early-career professionals.


## Cluster 7 Summary

for index, col in enumerate(df[categorical_features].columns):
    print("#"*32,col,"#"*32)
    print(df[df['Kmeans_labels'] == 7][col].value_counts(normalize=True) * 100, "\n")
    
    
for index, col in enumerate(df[continuous_features].columns):
    print("#"*32,col,"#"*32)
    print(f"{df[df['Kmeans_labels'] == 7][col].describe()}\n")


- **Education**: The education distribution in this cluster includes:
  - 54.15% have a Graduation degree, 14.80% hold a Master's degree, 13.72% have a PhD, 10.11% have a 2nd Cycle education, and 7.22% have a Basic education level.
- **Marital Status**: All individuals are "Uncommitted."
- **Family Structure**:
  - **Kidhome**: 77.62% have one child, 18.41% have no children, and 3.97% have two children.
  - **Teenhome**: 71.12% have no teenagers, 28.16% have one teenager, and 0.72% have two teenagers.
- **Campaign Engagement (Cmp_Accepted)**: A large proportion (91.34%) of individuals have not accepted any campaigns, with a small percentage (8.30%) accepting one campaign and 0.36% accepting two.

- **Tenure**: The average tenure in this cluster is 11.5 months, with a range from 0 to 22 months.
- **Age**: The average age is 38.53 years, with a range from 19 to 65 years.
- **Income**: On average, income is $31,121, ranging from $1,730 to $68,274.
- **Consumption**:
  - **Total Bought**: The average number of items purchased is 119, ranging from 5 to 835.
  - **Total Purchases**: On average, 8.52 purchases were made, with a range from 0 to 25.

### Key Insights
This cluster is characterized by individuals with a high proportion of people holding a Graduation education level (54.2%) and a relatively young to middle-aged group (average age of 38.5 years). Most individuals have one child, and a significant portion have no teenagers. Their tenure at the company is relatively moderate (average of 11.5 months). This group has an average income of 
$31,122, which is lower compared to other clusters, and shows low purchasing activity, with an average of 119 items bought. Only a small fraction has accepted campaigns, making them a potential target for specific offers, particularly for lower-cost or essential products.
